Update: car buying

+++ backend-api/controllers/management/OrdersController.js
@@ -1,13 +1,87 @@
 import { generatePaginationLinks } from "../../helper/PagingHelper.js";
 import Orders from "../../models/Orders.js";
 import { Op } from "sequelize";
+import Staff from "../../models/Staff.js";
+import Customer from "../../models/Customer.js";
+import Car from "../../models/Car.js";
+import OrderDetails from "../../models/OrderDetails.js";
+import Showroom from "../../models/Showroom.js";
+export const getOrder = async (req, res) => {
+  const { id } = req.params;
+  const staff = req.user;
+  if (isNaN(id)) {
+    return res.status(400).json({ error: "ID không hợp lệ" });
+  }
+  try {
+    const order = await Orders.findOne({
+      where: { id },
+      include: [
+        {
+          model: Staff,
+          as: "technical_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Staff,
+          as: "sales_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Staff,
+          as: "insurance_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Customer,
+          as: "customer",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Car,
+          as: "car",
+        },
+        {
+          model: Showroom,
+          as: "showroom",
+        },
+        {
+          model: OrderDetails,
+          as: "order_details",
+        },
+      ],
+    });
 
+    if (!order) {
+      return res.status(404).json({ error: "Yêu cầu không tồn tại" });
+    }
+    if (staff.role.name == "Sale") {
+      if (order.sales_staff_id != staff.id) {
+        return res.status(404).json({ error: "Yêu cầu không tồn tại" });
+      }
+    }
+    if (staff.role.name == "Insurance") {
+      if (order.insurance_staff_id != staff.id) {
+        return res.status(404).json({ error: "Yêu cầu không tồn tại" });
+      }
+    }
+    if (staff.role.name == "Technical") {
+      if (order.technical_staff_id != staff.id) {
+        return res.status(404).json({ error: "Yêu cầu không tồn tại" });
+      }
+    }
+    const data = order.toJSON();
+    return res.status(200).json({ data: data });
+  } catch (error) {
+    return res.status(500).json({ error: error.message ?? "Lỗi máy chủ" });
+  }
+};
 export const queryOrder = async (req, res) => {
   const perPage = parseInt(req.query.items_per_page) || 20;
   const currentPage = parseInt(req.query.page) || 1;
   const sortColumn = req.query.sort || "id";
   const sortOrder = req.query.order || "desc";
   const searchQuery = req.query.search || "";
+  const staff = req.user;
 
   try {
     const searchConditions = {};
@@ -40,7 +114,13 @@ export const queryOrder = async (req, res) => {
         }
       });
     }
-
+    if (staff.role.name == "Sale") {
+      searchConditions.sales_staff_id = staff.id;
+    } else if (staff.role.name == "Insurance") {
+      searchConditions.insurance_staff_id = staff.id;
+    } else if (staff.role.name == "Technical") {
+      searchConditions.technical_staff_id = staff.id;
+    }
     const totalOrders = await Orders.count({
       where: searchConditions,
     });
@@ -50,11 +130,34 @@ export const queryOrder = async (req, res) => {
       offset: (currentPage - 1) * perPage,
       limit: perPage,
       include: [
-        //   {
-        //     model: Brand,
-        //     as: "brand",
-        //     attributes: ["name"],
-        //   }
+        {
+          model: Staff,
+          as: "technical_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Staff,
+          as: "sales_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Staff,
+          as: "insurance_staff",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Customer,
+          as: "customer",
+          attributes: ["fullname", "email", "phone_number"],
+        },
+        {
+          model: Car,
+          as: "car",
+        },
+        {
+          model: Showroom,
+          as: "showroom",
+        },
       ],
       order: [[sortColumn, sortOrder.toUpperCase()]],
     });
@@ -103,58 +206,35 @@ export const queryOrder = async (req, res) => {
     res.status(500).json({ error: error.message || "Something went wrong" });
   }
 };
-
-export const getOrderById = async (req, res) => {
-  try {
-    const order = await Orders.findByPk(req.params.id, {
-      include: ["orderDetails", "car"],
-    });
-    if (order) {
-      res.status(200).json(order);
-    } else {
-      res.status(404).json({ message: "Order not found" });
-    }
-  } catch (error) {
-    res.status(500).json({ error: error.message });
+export const deleteOrders = async (req, res) => {
+  let Ids = req.body.ids;
+  if (!Ids || Ids.length === 0) {
+    res.status(500).json({ error: "Danh sách ID không hợp lệ" });
   }
-};
-
-export const createOrder = async (req, res) => {
-  try {
-    const newOrder = await Orders.create(req.body);
-    res.status(201).json(newOrder);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
+  Ids = Ids.filter((id) => !isNaN(id));
+  if (Ids.length === 0) {
+    return res.status(400).json({ error: "Không có ID hợp lệ để xóa" });
   }
-};
-
-export const updateOrder = async (req, res) => {
   try {
-    const updated = await Orders.update(req.body, {
-      where: { id: req.params.id },
+    const deletedDeltails = await OrderDetails.destroy({
+      where: {
+        order_id: Ids,
+      },
     });
-    if (updated) {
-      const updatedOrder = await Orders.findByPk(req.params.id);
-      res.status(200).json(updatedOrder);
-    } else {
-      res.status(404).json({ message: "Order not found" });
-    }
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-};
 
-export const deleteOrder = async (req, res) => {
-  try {
-    const deleted = await Orders.destroy({
-      where: { id: req.params.id },
+    const deletedCount = await Orders.destroy({
+      where: {
+        id: Ids,
+      },
     });
-    if (deleted) {
-      res.status(204).send("Order deleted");
-    } else {
-      res.status(404).json({ message: "Order not found" });
+
+    if (deletedCount === 0) {
+      return res.status(404).json({ error: "Không tìm thấy đơn hàng để xóa" });
     }
+
+    res.status(200).json({ error: "Xóa thành công", deletedCount });
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    console.error(error);
+    res.status(500).json({ error: "Lỗi máy chủ khi xóa đơn hàng" });
   }
 };
